barn: Preprocessing and Feature Engineering Steps before Modeling

The package provides pipeable functions to simplify preprocessing of tabular data prior to machine learning modeling. Users can combine multiple datasets, define feature engineering steps (such as creating new predictors from nominal or numeric columns), and then split the data back into preprocessed datasets ready to be used in machine learning workflows.

Author

Maintainer: Jordi Rosell jroselln@gmail.com (ORCID)